third_party/blink/web_tests/fast/table/table-rowspan-height-less-than-one-percent.html

<!DOCTYPE html>
<style>
    #perc-height, th { height: .5%; }
    table { height: 10000px }
</style>
<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('th');">
    <h3>Test for chromium bug : <a href="https://code.google.com/p/chromium/issues/detail?id=408066">408066</a>. ASSERTION FAILED: !extraRowSpanningHeight in blink::RenderTableSection::distributeRowSpanHeightToRows.</h3>
    <table border="1">
        <th id="perc-height" data-expected-height="50">
            <td rowspan="2">r0c0</td>
        </th>
        <tr>
        </tr>
        <tr>
            <td>r2c0</td>
        </tr>
    </table>
    <table border="1">
        <th id="perc-height" data-expected-height="50">
            <td>r0c0</td>
            <td rowspan="2">r0c1</td>
        </th>
        <tr>
        </tr>
        <tr>
            <td>r2c0</td>
        </tr>
    </table>
</body>
